run the linter and fix the linter errors

+++ dashboard/views.py
@@ -1,25 +1,26 @@
-from django.shortcuts import render, get_object_or_404, redirect
-from django.contrib.auth.decorators import login_required
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import get_object_or_404, redirect, render
 from django.utils import timezone
-from blog.models import Post
+
 from blog.forms import PostForm
+from blog.models import Post
 
 from .forms import ProfileForm
 
 # Create your views here.
 
 
-@login_required(login_url='/accounts/login')
+@login_required(login_url="/accounts/login")
 def index(request):
     posts = Post.objects.filter(owner = request.user)
-    return render(request, 'dashboard/index.html', {'posts': posts})
+    return render(request, "dashboard/index.html", {"posts": posts})
 
 
 
-@login_required(login_url='/accounts/login')
+@login_required(login_url="/accounts/login")
 def create_post(request):
-    if request.method == 'POST':
+    if request.method == "POST":
 
         form = PostForm(request.POST, request.FILES)
         if form.is_valid():
@@ -29,27 +30,26 @@ def create_post(request):
             post.save()
 
             form.save_m2m()
-            
-            messages.success(request, 'Post create successfully')
-            return redirect('/')
+
+            messages.success(request, "Post create successfully")
+            return redirect("/")
     else:
         form = PostForm()
-    return render(request, 'dashboard/post_create.html', {'form': form})
+    return render(request, "dashboard/post_create.html", {"form": form})
 
 
-@login_required(login_url='/accounts/login')
+@login_required(login_url="/accounts/login")
 def delete_post(request, slug):
     post = get_object_or_404(Post, slug = slug)
-    if request.method == "POST":
-        if request.user == post.owner:
-            messages.success(request, "the post deleted successfully")
-            post.delete()
-            return redirect('/dashboard')
-    return render(request, "dashboard/post_delete.html", {'post': post})
+    if request.method == "POST" and request.user == post.owner:
+        messages.success(request, "the post deleted successfully")
+        post.delete()
+        return redirect("/dashboard")
+    return render(request, "dashboard/post_delete.html", {"post": post})
 
-@login_required(login_url='/accounts/login')
+@login_required(login_url="/accounts/login")
 def update_post(request, slug):
-    
+
     post = get_object_or_404(Post, slug = slug)
     if request.user == post.owner:
         if request.method == "POST":
@@ -57,28 +57,28 @@ def update_post(request, slug):
             if form.is_valid():
                 form.save()
                 messages.success(request, "the post updated successfully")
-                return redirect('/dashboard')
+                return redirect("/dashboard")
     else:
         messages.warning(request, "this post is not yours for update" )
-        return redirect('/dashboard')
+        return redirect("/dashboard")
 
-    form = PostForm(instance=post)   
-    return render(request, "dashboard/update_post.html", {'form': form})
-    
+    form = PostForm(instance=post)
+    return render(request, "dashboard/update_post.html", {"form": form})
+
 
 
 
 
 @login_required
 def edit_profile(request):
-    if request.method == 'POST':
+    if request.method == "POST":
         form = ProfileForm(request.POST, instance=request.user.profile)
         if form.is_valid():
             form.save()
             messages.success(request, "Your profile update successfully")
-            return redirect('dashboard:edit_profile')
+            return redirect("dashboard:edit_profile")
 
     else:
         form = ProfileForm(instance=request.user.profile)
-    
-    return render(request, 'dashboard/profile.html', {'form': form})
+
+    return render(request, "dashboard/profile.html", {"form": form})
